requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Cython==0.29.16
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
jedi==0.17.0
parso==0.7.0
prompt-toolkit==3.0.5
ptpython==3.0.2
Pygments==2.6.1
pymssql==2.1.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.0.1
